First draft of the VM discovery UI - though it's displaying VMs, disks etc., it's still using the input file to get the specific VMs as oppose to listing all VMs in the target hosts.  Will fix this part later.  Will get the current process working through UI first.
Changed/added some fields to vms, e.g. cores=>src_cpus, memory=>src_memory so that existing index.tpl will work
Changed the input JSON and the 'vms' output and temp 'vms' variable to be dictionary instead of list so that we can make this hypervisor dependent codes as similar to the one for vmware, so that we can easily take out the common elements later.  instead of containing a list of vms, 'vms' now as hash vm id as key just like the output of vmware vms